+++ tests/testthat/test-rstanarm.R
@@ -5,6 +5,7 @@ if (.runThisTest) {
   if (suppressWarnings(
     require("testthat") &&
     require("lme4") &&
+    require("sjmisc") &&
     requireNamespace("rstanarm") &&
     require("ggeffects")
   )) {
@@ -12,8 +13,10 @@ if (.runThisTest) {
 
     # fit linear model
     data(sleepstudy)
+    data(efc)
     sleepstudy$age <- round(runif(nrow(sleepstudy), min = 20, max = 60))
     sleepstudy$Rdicho <- dicho(sleepstudy$Reaction)
+    efc <- to_label(efc, e42dep, c161sex, c172code)
 
     m <- rstanarm::stan_glmer(
       Reaction ~ Days + age + (1 | Subject),
@@ -26,6 +29,13 @@ if (.runThisTest) {
       Rdicho ~ Days + age + (1 | Subject),
       data = sleepstudy, QR = TRUE,
       family = binomial,
+      chains = 2, iter = 500
+    )
+
+    m3 <- rstanarm::stan_glm(
+      tot_sc_e ~ neg_c_7 + e42dep + barthtot + c172code + c161sex,
+      data = efc,
+      family = poisson("log"),
       chains = 2, iter = 500
     )
 
@@ -51,5 +61,11 @@ if (.runThisTest) {
       ggpredict(m2, c("Days", "age"), type = "re", ppd = TRUE)
     })
 
+    test_that("ggpredict, rstan", {
+      ggpredict(m3, "neg_c_7")
+      ggpredict(m3, c("neg_c_7", "e42dep"))
+      ggpredict(m3, "neg_c_7", ppd = TRUE)
+      ggpredict(m3, c("neg_c_7", "e42dep"), ppd = TRUE)
+    })
   }
 }